+++ cmd/moombox/main.go
@@ -45,7 +45,7 @@ import (
 )
 
 var (
-	version = "2.3.17"
+	version = "2.3.18"
 	commit  = ""
 )